Time: 102 ms (35.35%), Space: 125 MB (15.18%) - LeetHub

+++ 0219-contains-duplicate-ii/0219-contains-duplicate-ii.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool containsNearbyDuplicate(vector<int>& nums, int k) {
         unordered_map<int, int> lastSeen; // value -> most recent index
-        for (int i = 0; i < (int)nums.size(); i++) {
+        for (int i = 0; i < nums.size(); i++) {
             auto it = lastSeen.find(nums[i]);
             if (it != lastSeen.end() && i - it->second <= k) {
                 return true;
